Basic Auth Controller
Added Basic Auth Controller for HTTP Basic Auth setup. Extending OAuth
from this, and altered how the Bearer Token is configured, by using a
method that can be overridden in any class that extends it.

Changed DEFAULT AUTH ACTIONS to be protected, so that it can be
overridden and defined by extended classes.